v1.6.5

+++ include/ota.h
@@ -17,7 +17,7 @@ int version_compare(const String v1, const String v2);
 void display(const uint8_t row, const std::string status,
                        const std::string msg);
 #ifdef HAS_DISPLAY
-void show_progress(size_t current, size_t size);
+void show_progress(unsigned long current, unsigned long size);
 #endif
 
 #endif // USE_OTA
